Changed Files Export
There are directions at the top to assume what a server name is.

Also when we mount the filesystem, we mount the share and not the root filesystem. The old document assumes the student is making a filesystem for their own use, but this lab, needs to assume the filesystem is already created and thus, activity needs to be driven to the share, not the root.